+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # develop
+  * Bugfix: The Asterisk translator now ignores calls to 'h' or of type 'Kill'
 
 # v0.9.2 - 2012-02-18
   * Feature: Asterisk calls receiving media commands are implicitly answered